add script for regenerating protobuf files and update files
