--- src/protobuf/akash/base/v1beta2/resource.ts
@@ -1,304 +1,0 @@
-/* eslint-disable */
-import Long from "long";
-import _m0 from "protobufjs/minimal";
-import { ResourceValue } from "../../../akash/base/v1beta2/resourcevalue";
-import { Attribute } from "../../../akash/base/v1beta2/attribute";
-
-export const protobufPackage = "akash.base.v1beta2";
-
-/** CPU stores resource units and cpu config attributes */
-export interface CPU {
-  units?: ResourceValue;
-  attributes: Attribute[];
-}
-
-/** Memory stores resource quantity and memory attributes */
-export interface Memory {
-  quantity?: ResourceValue;
-  attributes: Attribute[];
-}
-
-/** Storage stores resource quantity and storage attributes */
-export interface Storage {
-  name: string;
-  quantity?: ResourceValue;
-  attributes: Attribute[];
-}
-
-function createBaseCPU(): CPU {
-  return { units: undefined, attributes: [] };
-}
-
-export const CPU = {
-  encode(message: CPU, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
-    if (message.units !== undefined) {
-      ResourceValue.encode(message.units, writer.uint32(10).fork()).ldelim();
-    }
-    for (const v of message.attributes) {
-      Attribute.encode(v!, writer.uint32(18).fork()).ldelim();
-    }
-    return writer;
-  },
-
-  decode(input: _m0.Reader | Uint8Array, length?: number): CPU {
-    const reader = input instanceof _m0.Reader ? input : new _m0.Reader(input);
-    let end = length === undefined ? reader.len : reader.pos + length;
-    const message = createBaseCPU();
-    while (reader.pos < end) {
-      const tag = reader.uint32();
-      switch (tag >>> 3) {
-        case 1:
-          message.units = ResourceValue.decode(reader, reader.uint32());
-          break;
-        case 2:
-          message.attributes.push(Attribute.decode(reader, reader.uint32()));
-          break;
-        default:
-          reader.skipType(tag & 7);
-          break;
-      }
-    }
-    return message;
-  },
-
-  fromJSON(object: any): CPU {
-    return {
-      units: isSet(object.units)
-        ? ResourceValue.fromJSON(object.units)
-        : undefined,
-      attributes: Array.isArray(object?.attributes)
-        ? object.attributes.map((e: any) => Attribute.fromJSON(e))
-        : [],
-    };
-  },
-
-  toJSON(message: CPU): unknown {
-    const obj: any = {};
-    message.units !== undefined &&
-      (obj.units = message.units
-        ? ResourceValue.toJSON(message.units)
-        : undefined);
-    if (message.attributes) {
-      obj.attributes = message.attributes.map((e) =>
-        e ? Attribute.toJSON(e) : undefined
-      );
-    } else {
-      obj.attributes = [];
-    }
-    return obj;
-  },
-
-  fromPartial<I extends Exact<DeepPartial<CPU>, I>>(object: I): CPU {
-    const message = createBaseCPU();
-    message.units =
-      object.units !== undefined && object.units !== null
-        ? ResourceValue.fromPartial(object.units)
-        : undefined;
-    message.attributes =
-      object.attributes?.map((e) => Attribute.fromPartial(e)) || [];
-    return message;
-  },
-};
-
-function createBaseMemory(): Memory {
-  return { quantity: undefined, attributes: [] };
-}
-
-export const Memory = {
-  encode(
-    message: Memory,
-    writer: _m0.Writer = _m0.Writer.create()
-  ): _m0.Writer {
-    if (message.quantity !== undefined) {
-      ResourceValue.encode(message.quantity, writer.uint32(10).fork()).ldelim();
-    }
-    for (const v of message.attributes) {
-      Attribute.encode(v!, writer.uint32(18).fork()).ldelim();
-    }
-    return writer;
-  },
-
-  decode(input: _m0.Reader | Uint8Array, length?: number): Memory {
-    const reader = input instanceof _m0.Reader ? input : new _m0.Reader(input);
-    let end = length === undefined ? reader.len : reader.pos + length;
-    const message = createBaseMemory();
-    while (reader.pos < end) {
-      const tag = reader.uint32();
-      switch (tag >>> 3) {
-        case 1:
-          message.quantity = ResourceValue.decode(reader, reader.uint32());
-          break;
-        case 2:
-          message.attributes.push(Attribute.decode(reader, reader.uint32()));
-          break;
-        default:
-          reader.skipType(tag & 7);
-          break;
-      }
-    }
-    return message;
-  },
-
-  fromJSON(object: any): Memory {
-    return {
-      quantity: isSet(object.quantity)
-        ? ResourceValue.fromJSON(object.quantity)
-        : undefined,
-      attributes: Array.isArray(object?.attributes)
-        ? object.attributes.map((e: any) => Attribute.fromJSON(e))
-        : [],
-    };
-  },
-
-  toJSON(message: Memory): unknown {
-    const obj: any = {};
-    message.quantity !== undefined &&
-      (obj.quantity = message.quantity
-        ? ResourceValue.toJSON(message.quantity)
-        : undefined);
-    if (message.attributes) {
-      obj.attributes = message.attributes.map((e) =>
-        e ? Attribute.toJSON(e) : undefined
-      );
-    } else {
-      obj.attributes = [];
-    }
-    return obj;
-  },
-
-  fromPartial<I extends Exact<DeepPartial<Memory>, I>>(object: I): Memory {
-    const message = createBaseMemory();
-    message.quantity =
-      object.quantity !== undefined && object.quantity !== null
-        ? ResourceValue.fromPartial(object.quantity)
-        : undefined;
-    message.attributes =
-      object.attributes?.map((e) => Attribute.fromPartial(e)) || [];
-    return message;
-  },
-};
-
-function createBaseStorage(): Storage {
-  return { name: "", quantity: undefined, attributes: [] };
-}
-
-export const Storage = {
-  encode(
-    message: Storage,
-    writer: _m0.Writer = _m0.Writer.create()
-  ): _m0.Writer {
-    if (message.name !== "") {
-      writer.uint32(10).string(message.name);
-    }
-    if (message.quantity !== undefined) {
-      ResourceValue.encode(message.quantity, writer.uint32(18).fork()).ldelim();
-    }
-    for (const v of message.attributes) {
-      Attribute.encode(v!, writer.uint32(26).fork()).ldelim();
-    }
-    return writer;
-  },
-
-  decode(input: _m0.Reader | Uint8Array, length?: number): Storage {
-    const reader = input instanceof _m0.Reader ? input : new _m0.Reader(input);
-    let end = length === undefined ? reader.len : reader.pos + length;
-    const message = createBaseStorage();
-    while (reader.pos < end) {
-      const tag = reader.uint32();
-      switch (tag >>> 3) {
-        case 1:
-          message.name = reader.string();
-          break;
-        case 2:
-          message.quantity = ResourceValue.decode(reader, reader.uint32());
-          break;
-        case 3:
-          message.attributes.push(Attribute.decode(reader, reader.uint32()));
-          break;
-        default:
-          reader.skipType(tag & 7);
-          break;
-      }
-    }
-    return message;
-  },
-
-  fromJSON(object: any): Storage {
-    return {
-      name: isSet(object.name) ? String(object.name) : "",
-      quantity: isSet(object.quantity)
-        ? ResourceValue.fromJSON(object.quantity)
-        : undefined,
-      attributes: Array.isArray(object?.attributes)
-        ? object.attributes.map((e: any) => Attribute.fromJSON(e))
-        : [],
-    };
-  },
-
-  toJSON(message: Storage): unknown {
-    const obj: any = {};
-    message.name !== undefined && (obj.name = message.name);
-    message.quantity !== undefined &&
-      (obj.quantity = message.quantity
-        ? ResourceValue.toJSON(message.quantity)
-        : undefined);
-    if (message.attributes) {
-      obj.attributes = message.attributes.map((e) =>
-        e ? Attribute.toJSON(e) : undefined
-      );
-    } else {
-      obj.attributes = [];
-    }
-    return obj;
-  },
-
-  fromPartial<I extends Exact<DeepPartial<Storage>, I>>(object: I): Storage {
-    const message = createBaseStorage();
-    message.name = object.name ?? "";
-    message.quantity =
-      object.quantity !== undefined && object.quantity !== null
-        ? ResourceValue.fromPartial(object.quantity)
-        : undefined;
-    message.attributes =
-      object.attributes?.map((e) => Attribute.fromPartial(e)) || [];
-    return message;
-  },
-};
-
-type Builtin =
-  | Date
-  | Function
-  | Uint8Array
-  | string
-  | number
-  | boolean
-  | undefined;
-
-export type DeepPartial<T> = T extends Builtin
-  ? T
-  : T extends Long
-  ? string | number | Long
-  : T extends Array<infer U>
-  ? Array<DeepPartial<U>>
-  : T extends ReadonlyArray<infer U>
-  ? ReadonlyArray<DeepPartial<U>>
-  : T extends {}
-  ? { [K in keyof T]?: DeepPartial<T[K]> }
-  : Partial<T>;
-
-type KeysOfUnion<T> = T extends T ? keyof T : never;
-export type Exact<P, I extends P> = P extends Builtin
-  ? P
-  : P & { [K in keyof P]: Exact<P[K], I[K]> } & Record<
-        Exclude<keyof I, KeysOfUnion<P>>,
-        never
-      >;
-
-if (_m0.util.Long !== Long) {
-  _m0.util.Long = Long as any;
-  _m0.configure();
-}
-
-function isSet(value: any): boolean {
-  return value !== null && value !== undefined;
-}
